Styling
In this commit I added styling to make the placeholder look somewhat
decent. I am using a google font, to have the text be a bit more
remarkable and the CSS is mostly normalize.css to have things look
coherent in all browsers.
the text was just positioned in the center, the title got an underline
and the link color was adapted.