Preparation for GitLab Pages
In this commit I restructured the project to use GitHub Pages instead of
a managed hosting solution. This is because the managed hosting requires
a lot of overhead to setup, especially considering emails, so I don't
want to bother with it anymore. In the search for alternative mail
solutions I openend my mind and found that I also could host the website
easier and especially cheaper. So this is exactly what I am trying now.
I hope this works as expected, with SSL, redirects and automatic deploy.
The latter is quite nice as it seems quite easy to get going. Nothing
like a daploy per FTP.